refactor(lib): rename storage.ts -> sessionRepository.ts (Supabase DB repo, not browser storage)
storage.ts is the client-side repository for persisting/reading practice sessions,
profiles, and history from Supabase — it does NOT touch localStorage/sessionStorage
(that is safeStorage.ts). The name caused recurring semantic confusion (codebase-audit
finding BL-2). Pure rename: git mv + rewrite of every importer/mock/dynamic-import
specifier; 2 dedicated test files renamed to match. No behavioral change.

Verified: tsc clean; affected importer suites 64/64; repository tests 14/14; eslint clean.

Done on the dev/v4-integration line (where the heavily-edited importers SpeechRuntimeController
and TranscriptionService currently live) to avoid a main<->v4 import-line merge conflict;
supersedes the interim clarifying header on #756.
